fix location of email settings types
